fix db

+++ database/migrations/2025_06_10_180303_create_payments_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('payments', function (Blueprint $table) {
             $table->id('payment_id')->comment('支払情報ID');
             $table->string('payment_name')->comment('支払方法名');
-            $table->integer('card_number')->nullable()->comment('カード番号');
+            $table->bigInteger('card_number')->nullable()->comment('カード番号');
             $table->integer('expire')->nullable()->comment('有効期限');
             $table->string('card_customer_name')->nullable()->comment('カード名義');
             $table->boolean('can_use_flg')->comment('有効な支払方法フラグ');
